+++ includes/camera_info.cuh
@@ -30,8 +30,8 @@ enum class CAMERA_MODEL {
 // TODO: Check and remove this struct if possible
 struct CameraInfo {
     uint32_t _camera_ID;
-    Eigen::Matrix3d _R; // rotation  matrix
-    Eigen::Vector3d _T; // translation vector
+    Eigen::Matrix3f _R; // rotation  matrix
+    Eigen::Vector3f _T; // translation vector
     float _fov_x;
     float _fov_y;
     std::string _image_name;
@@ -65,19 +65,19 @@ inline void dump_JSON(const std::filesystem::path& file_path, std::vector<nlohma
 }
 
 // serialize camera to json
-inline nlohmann::json Convert_camera_to_JSON(const CameraInfo& cam, int id, Eigen::Matrix3d& R, const Eigen::Vector3d& T) {
+inline nlohmann::json Convert_camera_to_JSON(const CameraInfo& cam, int id, Eigen::Matrix3f& R, const Eigen::Vector3f& T) {
 
-    Eigen::Matrix4d Rt = Eigen::Matrix4d::Zero();
+    Eigen::Matrix4f Rt = Eigen::Matrix4f::Zero();
     Rt.block<3, 3>(0, 0) = R.transpose();
     Rt.block<3, 1>(0, 3) = T;
     Rt(3, 3) = 1.0;
 
-    Eigen::Matrix4d W2C = Rt.inverse();
-    Eigen::Vector3d pos = W2C.block<3, 1>(0, 3);
-    Eigen::Matrix3d rot = W2C.block<3, 3>(0, 0);
-    std::vector<std::vector<double>> serializable_array_2d;
+    Eigen::Matrix4f W2C = Rt.inverse();
+    Eigen::Vector3f pos = W2C.block<3, 1>(0, 3);
+    Eigen::Matrix3f rot = W2C.block<3, 3>(0, 0);
+    std::vector<std::vector<float>> serializable_array_2d;
     for (int i = 0; i < rot.rows(); i++) {
-        serializable_array_2d.push_back(std::vector<double>(rot.row(i).data(), rot.row(i).data() + rot.row(i).size()));
+        serializable_array_2d.push_back(std::vector<float>(rot.row(i).data(), rot.row(i).data() + rot.row(i).size()));
     }
 
     nlohmann::json camera_entry = {
@@ -85,7 +85,7 @@ inline nlohmann::json Convert_camera_to_JSON(const CameraInfo& cam, int id, Eige
         {"img_name", cam._image_name},
         {"width", cam._width},
         {"height", cam._height},
-        {"position", std::vector<double>(pos.data(), pos.data() + pos.size())},
+        {"position", std::vector<float>(pos.data(), pos.data() + pos.size())},
         {"rotation", serializable_array_2d},
         {"fy", fov2focal(cam._fov_y, cam._height)},
         {"fx", fov2focal(cam._fov_x, cam._width)}};
